% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\docType{package}
\encoding{UTF-8}
\name{measurements}
\alias{measurements}
\title{Tools for Units of Measurement}
\description{
Collection of tools to make working with physical measurements easier. Convert between metric and imperial units, or calculate a dimension's unknown value from other dimensions' measurements.
}
\author{
Matthew A. Birk, \email{matthewabirk@gmail.com}
}
